+++ src/main/java/de/stamme/basicquests/model/rewards/RewardType.java
@@ -1,7 +1,21 @@
 package de.stamme.basicquests.model.rewards;
 
+import de.stamme.basicquests.config.Config;
+
 public enum RewardType {
     ITEM,
     MONEY,
-    XP
+    XP;
+
+    public double getWeight() {
+        switch (this) {
+            case ITEM:
+                return Config.getItemRewardsWeight();
+            case MONEY:
+                return Config.getMoneyRewardsWeight();
+            case XP:
+                return Config.getXpRewardsWeight();
+        }
+        return 1.0;
+    }
 }